refactor(architecture): separate repositories from services and simplify module boundaries
Move env/git repository implementations into src/repositories and keep src/services focused on orchestration logic. Also remove barrel wrappers and legacy service entry files so dependencies are imported from explicit module paths.
