+++ src/components/AgentLoop.tsx
@@ -3,7 +3,8 @@ import type z from "zod";
 import { useAppDependencies } from "../contexts/app_dependencies_context.tsx";
 // import { renderTui } from "../lib/opentui_render.tsx";
 import { IssueAgentSchema, type IssueSchema } from "../schema.ts";
-import { AIService, type UsageCallback } from "../services/ai.ts";
+import { AIService } from "../services/ai/ai_service.ts";
+import type { UsageCallback } from "../services/ai/types.ts";
 import { Box } from "./ThemedComponents.tsx";
 import { Spinner } from "./Spinner.tsx";
 import { TextInput } from "./TextInput.tsx";
